Remove setup of test image with docker and golem binary
The base image should contain the docker and golem binary needed to run the test.
The base image must have a golem binary with the name "golem_runner" which will be invoked by the test runner.
Docker binary and version management is no longer done by the runner but relies on using a base image with the correct docker version.
The golem runner now requires docker 1.10 to run, any version of docker may be run in the test containers but 1.10 should be used by the test base image.

Fixes #1
Fixes #3

Signed-off-by: Derek McGowan <derek@mcgstyle.net> (github: dmcgowan)